+++ src/master/master/setUp.go
@@ -8,7 +8,6 @@ import (
 	"net/http"
 	"fmt"
 	"flag"
-	"os/exec"
 )
 
 type Slave struct {
@@ -16,22 +15,6 @@ type Slave struct {
 	Heartbeat              time.Time
 	PreviouslyDisplayedURL string
 	DisplayedURL           string
-}
-
-func FlushIPTables() (error error) {
-	error = exec.Command("sudo", "iptables", "--policy", "INPUT", "DROP").Run()
-	if error != nil {
-		fmt.Printf("Error flushing iptables: %v\n", error)
-	}
-	return
-}
-
-func AcceptResponseFromDNSServer() (error error) {
-	error = exec.Command("sudo", "iptables", "-A", "INPUT", "-m", "conntrack", "--ctstate", "RELATED,ESTABLISHED", "-j", "ACCEPT").Run()
-	if error != nil {
-		fmt.Printf("Error setting rule for accepting responses from DNS server: %v\n", error)
-	}
-	return
 }
 
 func GetSlaveMap() (slaveMap map[string]Slave) {
